// microservice_erp/microservice-upload/src/modules/item-print/interfaces/response.ts
export interface MetadataResponse {
    success: boolean;
    message: string;
    data: string;

}
export interface MetadataP2Response {
    success: boolean;
    message: string;
    data: string;
    maintLogDetails: string;

}
export interface MetadataResponse2 {
    success: boolean;
    message: string;
    data: string;
    errors?: any[];

}